Create a base class for inputs. Most inputs have mostly the same behavior. Generally we will simply have to extend the component options with a template and any additional bindings that are needed.
