+++ source/components/components.module.ts
@@ -15,6 +15,7 @@ import * as dateTime from './dateTime/dateTime';
 import * as dateTimeStatic from './dateTimeStatic/dateTimeStatic';
 import * as dialog from './dialog/dialog';
 import * as genericContainer from './genericContainer/genericContainer';
+import * as input from './input/input';
 import * as lazyLoad from './lazyLoad/lazyLoad';
 import * as longClickButton from './longClickButton/longClickButton';
 import * as messageLog from './messageLog/messageLog.module';
